requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.3
python-dotenv==1.0.0
Werkzeug==2.2.3
Flask-Cors==3.0.10
Flask-Bcrypt==1.0.1
gunicorn
psycopg2-binary
Flask-Migrate